refactor: movieLoader 함수 파일분리
- render.js 생성

- skeleton.js 생성
